requirements.txt
fastapi
fastapi-utilities
httpx
uvicorn
icalendar
pyyaml